I'm getting closer to the source of the strange read errors
Probably found it actually.

I believe it would have been an issue with invalid references and whatnot.

So I made the shared references shared pointers.

Also instead of storing a raw pointer to the dialogue command I'm using a weak pointer now because I believe that the pointer becomes invalid. Possibly due to PIE mode. But all in all: Everyone warns you about using raw pointers without UPROPERTYs and if they can't be used then I guess this is the way to go.