Starting on remaking the dialoged language
Beginning with a simple Tokenizer

It's messy with the precedences but it's 1000% better than the current parser.

This will also change the syntax a little bit and, I guess, make it somewhat more fluent in terms of logic.

But me being me, had to add "" around text bodys to make it work.
Bit more typing involved now but I think it's still okay.