Commiting this cause I really don't want to lose it
This is the start of the AST paser that takes the tokens and turns them into an AST.

It's rough and missing some pieces that I will work on.
Once it's in a shape that handles all the features
and actually works, I'll come up with a more elegant solution to building it.

It's better than the old text parser for sure.
But it's too spaghetti for my liking