Added choices to the ast
Also removed the threadnode as it did not bring any meaning.

I might revise that later but I oughta keep it simple for the time being.

Crazy enough as it is

I also wanted to implement pre/post inc/dec operations but I'll leave this for now because I want to move on and I'm okay with unrolling that manually to a full assignment.