Making nodes uobjects
For several reasons.
But most importantly, They can be used now to just BE the actual execution stack, there is probably no need to convert them into more concrete data still.

Also preliminary checks seem to imply that this is working correctly in terms of how the tree is assembled.

But I need to check further still.