Removing something
The refactoring is done for now.
There are still some things missing that need to come back.

The flag that a branch has been run at least once, for instance.
This is not so easy anymore though because the are not distinct threads anymore and I need to find another way to address them

And then of course the pool of all actors within any given story thread.
The purpose that it used to fulfil was to facilitate the speech bubble tech on screen.
The code for which actually lies in a private repo.
That begs the question if there is a requirement to actually have this here or not.

I'll cross that bridge when I get back there.